replace int with long

+++ psets/7-basic-types/factorial.c
@@ -2,15 +2,15 @@
 
 int main(void)
 {
-	int n, fact;
+	long n, fact;
 	printf("Enter a positive integer: ");
-	scanf("%d", &n);
+	scanf("%ld", &n);
 	fact = n;
 	for (int i = 1; i < n; i++)
 	{
 		fact *= (n - i);
 	}
-	printf("Factorial of %d: %d\n", n, fact);
+	printf("Factorial of %ld: %ld\n", n, fact);
 }
 // !6 = 6 * 5 * 4 * 3 * 2 * 1
 // 		= 30 * 4 * 3 * 2 * 1
